Work on stereo configs
It turns out that even with 5.1 source material, ChromeOS will downmix
to stereo so early that we do not even get to process the 5.1 sound
through the filter graph.  This may not be true on another OS, though,
so I will map out the surround sound at some other time.