Fix filters when disabling and re-enabling UpFish
Two things were going wrong when disabling and re-enabling UpFish.

First, when disabling UpFish, there was an error thrown from the karaoke
worklet.  This seems to be caused by disconnecting the node, and could
be safely ignored.  However, ignoring errors is a bad habit that could
make it harder to spot an important error later.  Since disconnecting
the node is expected, we should handle that condition and not let the
worklet throw.

Second, when re-enabling UpFish, the karaoke filter was effectively
bypassed.  This is because when we disable UpFish, we can't explicitly
destroy any of our nodes.  Instead, we have to disconnect the filters,
and reconnect the source directly to the output.  If, on re-enabling
UpFish, we don't disconnect the source from the output again, the source
is effectively split.  It both passes through the filters and is also
copied to the output around the filters.  The solution is to simply
disconnect the source again when we re-enable UpFish.