Simplify rollup
Instead of rolling up in ES format, then editing the result, we can
simplify things.  By switching to default export instead of named
export for UpFish, we can make IIFE work correctly.  The previous way
of doing things was just a result of my lack of experience with rollup
and modules.